Small changes to support reading CommonCrawl files from S3
This change makes a few modifications to the HDFS utils. Importantly,
the `FileSystem` objects from the hadoop libraries are retrieved from
the URI of the files. This will allow accessing CommonCrawl WARC files
on filesystems other than the currently configured one in the HadoopConf.

Additionally there is a small fix for some sometimes corrupted WARC records
encountered in the output from CommonCrawl.